fix DBContentRepositoryTests on tmpfs
Introduce new configuration option
nifi.database.content.repository.use.direct.io with default value true,
overridden to false in tests.

The reason of failure on tmpfs is that tmpfs doesn't support Direct I/O,
but we try to use it with rocksdb, even in our tests that put the test
repos in /tmp. /tmp is often mounted as tmpfs, causing the tests to
fail.